INDY-1876: Change unstashing for view change and add tests
